discovery-api/requirements.txt
flask==3.1.0
flask-cors
redis>=6.0.0
ray[default]==2.46.0
python-dotenv
